requirements.txt
numpy
libigl
polyscope
ipctk
torch
seaborn
tqdm
ipykernel